Database lab

+++ src/test/java/test/pivotal/pal/tracker/InMemoryTimeEntryRepositoryTest.java
@@ -73,7 +73,7 @@ public class InMemoryTimeEntryRepositoryTest {
                 created.getId(),
                 new TimeEntry(321L, 654L, LocalDate.parse("2017-01-09"), 5));
 
-        TimeEntry expected = new TimeEntry((Long) created.getId(), 321L, 654L, LocalDate.parse("2017-01-09"), 5);
+        TimeEntry expected = new TimeEntry(created.getId(), 321L, 654L, LocalDate.parse("2017-01-09"), 5);
         assertThat(updatedEntry).isEqualTo(expected);
         assertThat(repo.find(created.getId())).isEqualTo(expected);
     }
